requirements.txt
# Project Pluto - Python Dependencies

# Core speech and audio (Whisper for better accuracy!)
openai-whisper>=20231117
pyaudio>=0.2.13

# LLM API
requests>=2.31.0

# System monitoring
psutil>=5.9.6

# Data handling
numpy>=1.24.0

# Optional: For advanced audio processing
scipy>=1.11.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
